+++ NEWS.md
@@ -10,6 +10,8 @@
 * EPW `COMMENT1` and `COMMENT2` are now parsed as a single string (#318).
 * Additional columns in `Epw$set()` and `Epw$add()` input are now removed
   (#320).
+* Now `Epw$abnormal_data()` can keep columns that contains abnormal data when
+  `keep_all` is `FALSE` (#326).
 
 # eplusr 0.13.0